# api/swagger.yml
openapi: 3.0.0
info:
  title: Market Performance API
  description: API to analyze market trends, track sales rank performance over time, and provide best performing products by category.
  version: 1.0.0
servers:
  - url: /api/v1
paths:
  /metadata:
    get:
      summary: Retrieve product metadata
      operationId: metadata_routes.get_metadata
      parameters:
        - in: query
          name: category
          required: true
          schema:
            type: string
        - in: query
          name: brand
          schema:
            type: string
        - in: query
          name: minPrice
          schema:
            type: number
            format: float
        - in: query
          name: maxPrice
          schema:
            type: number
            format: float
      responses:
        '200':
          description: A list of product metadata matching the filter criteria.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'
  /categories:
    get:
      summary: Retrieve main categories
      operationId: metadata_routes.get_main_categories
      responses:
        '200':
          description: A list of main categories available.
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
  /top-products:
    get:
      summary: Retrieve top ranking products in a category
      operationId: metadata_routes.get_top_products
      parameters:
        - in: query
          name: category
          required: true
          schema:
            type: string
      responses:
        '200':
          description: A list of top ranking products in the specified category.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'
  /top-brand:
    get:
      summary: Retrieve the top brand in a category
      operationId: metadata_routes.get_top_brand
      parameters:
        - in: query
          name: category
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The top brand in the specified category and the count of products.
          content:
            application/json:
              schema:
                type: object
                properties:
                  top_brand:
                    type: string
                  count:
                    type: integer

components:
  schemas:
    Product:
      type: object
      properties:
        title:
          type: string
        brand:
          type: string
        price:
          type: number
          format: float
        rank:
          type: integer
        categories:
          type: array
          items:
            type: string